backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
httpx==0.27.0
google-generativeai==0.5.4
pydantic==2.7.1
pydantic-core==2.18.2
pydantic-settings==2.2.1
python-dotenv==1.0.1
slowapi==0.1.9
structlog==24.1.0
firebase-admin==6.5.0
cachetools==5.3.3
beautifulsoup4==4.12.3
lxml==5.2.1
pdfplumber==0.11.0
sqlalchemy==2.0.29
nest_asyncio==1.6.0
tenacity==8.2.3
aiosqlite==0.20.0
